wip: style(GC4tA): tidy group A remainder — test narrowing, std::numbers, element-wise vec3 compare, MeshBevel widening, nested ?:
Also drops the int32 round trip on MergeCoincidentMeshEdges' float priority and
value-initialises two split/UV records. Editor builds (make -k); the 21 group-A
suites pass with unchanged counts.

+++ Desert/Desert/Source/Engine/Geometry/UECore/DynamicMesh/Operations/MergeCoincidentMeshEdges.cpp
@@ -162,7 +162,7 @@ namespace Desert::Geometry
                     continue;
             }
             DuplicatesQueue.Insert( eid,
-                                    static_cast<int32_t>( static_cast<float>( EquivalenceSets[eid]->size() ) ) );
+                                    static_cast<float>( EquivalenceSets[eid]->size() ) );
         }
 
         // greedy merge
